+++ packages/vue-dot/src/rules/notAfterToday/index.ts
@@ -3,24 +3,22 @@ import { ValidationRule, ValidationResult, ErrorMessages, Value } from '../types
 
 import { defaultErrorMessages } from './locales';
 
-import { isDateBeforeNow } from './isDateBeforeNow';
+import { isDateAfterValue } from '../../functions/isDateAfterValue';
+import { TODAY } from '../../constants';
 
 /**
- * Check that the date is not after today (expects ##/##/#### format)
+ * Check that the date is not after today (expects DD/MM/YYYY format)
  *
  * @param {ErrorMessages} [errorMessages] Custom error messages
  * @returns {ValidationRule} Validation result
  */
 export function notAfterTodayFn(errorMessages: ErrorMessages = defaultErrorMessages): ValidationRule {
 	return (value: Value): ValidationResult => {
-		// If the value is empty, return true (valid)
 		if (!value) {
 			return true;
 		}
 
-		// If the date is before now, it's a past date, it's valid,
-		// else, the date is after today, it's invalid
-		return isDateBeforeNow(value) || ruleMessage(errorMessages, 'default');
+		return !isDateAfterValue(TODAY, value) || ruleMessage(errorMessages, 'default');
 	};
 }